+++ routers/settings.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter
-from datasource import datasource
+from data import datasource
 from models.aroio import *
 
 ############
